Establish ownership boundaries between Discord message-turns and progress-leases modules
